feature(spaceOptionRepository): defined
