+++ packages/app/src/api/repositories/index.ts
@@ -6,5 +6,5 @@ export * from './userRepository';
 export * from './userProfileRepository';
 export * from './pluginsRepository';
 export * from './spaceAttributeRepository';
-export * from './spaceRepository';
+export * from './spaceOptionRepository';
 export * from './userAttributeRepository';
