+++ packages/app/src/api/repositories/spaceOptionRepository/index.ts
@@ -1,0 +1,1 @@
+export * from './spaceOptionRepository.api';
