+++ packages/app/src/api/repositories/spaceOptionRepository/spaceOptionRepository.api.types.ts
@@ -1,0 +1,16 @@
+import {OptionsInterface} from 'api/interfaces';
+
+export interface SpaceOptionRequest {
+  worldId: string;
+  spaceId: string;
+}
+
+// GET SPACE OPTIONS
+export interface GetSpaceOptionsRequest extends SpaceOptionRequest {}
+
+export interface GetSpaceOptionsResponse extends OptionsInterface {}
+
+// GET SPACE SUBOPTION
+export interface GetSpaceSubOptionRequest extends SpaceOptionRequest {
+  sub_option_key: string;
+}
